requirements.txt
# Core dependencies
yfinance>=0.2.36
pandas>=2.0.0
numpy>=1.24.0
openpyxl>=3.1.0
sqlalchemy>=2.0.0
pyyaml>=6.0
fastapi>=0.109.0
uvicorn>=0.27.0
anthropic>=0.18.0
openai>=1.12.0
python-dotenv>=1.0.0

# Development dependencies
pytest>=8.0.0
pytest-asyncio>=0.23.0
